Move messaging-contacts-bridge to setup layer
Bridge is orchestration/cross-domain wiring, not feature logic.
Relocate to src/setup/ alongside other composition patterns
(setupAuth, setupMainAppBusListeners).

Solves feature boundary violation: setup layer can import features,
features cannot import each other.
