+++ src/setup/tests/messaging-contacts-bridge.test.js
@@ -14,14 +14,14 @@ const mocks = vi.hoisted(() => {
   };
 });
 
-vi.mock('./messaging-controller.js', () => ({
+vi.mock('../../features/messaging/messaging-controller.js', () => ({
   messagingController: {
     getConversation: mocks.getConversation,
     on: mocks.on,
   },
 }));
 
-vi.mock('../contacts/index.js', () => ({
+vi.mock('../../features/contacts/index.js', () => ({
   contactsService: {
     updateLastInteraction: mocks.updateLastInteraction,
   },
@@ -37,7 +37,7 @@ describe('setupMessagingContactsIntegration', () => {
 
   it('updates the matching contact when messaging events arrive', async () => {
     const { setupMessagingContactsIntegration } = await import(
-      './messaging-contacts-bridge.js'
+      '../messaging-contacts-bridge.js'
     );
     const teardown = setupMessagingContactsIntegration();
 
@@ -53,7 +53,7 @@ describe('setupMessagingContactsIntegration', () => {
 
   it('ignores non-1:1 conversations', async () => {
     const { setupMessagingContactsIntegration } = await import(
-      './messaging-contacts-bridge.js'
+      '../messaging-contacts-bridge.js'
     );
     const teardown = setupMessagingContactsIntegration();
